% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedViolins.R
\name{groupedViolins}
\alias{groupedViolins}
\title{groupedViolins}
\usage{
groupedViolins(object, genes = NULL, marker_list = NULL,
  group_var = NULL, plot_title = NULL, assay_use = "RNA",
  flip = TRUE)
}
\arguments{
\item{object}{Seurat object to plot}

\item{genes}{List of genes to plot}

\item{marker_list}{A table of cluster markers as produced by FindMarkers()
or FindAllMarkers or a two column data.frame with the columns "cluster" and
"gene".  If provided, supercedes the genes argument.  Note: if a marker is
present in multiple clusters, only the first (as determined by arrange()) is used.
Default: NULL}

\item{group_var}{Grouping variable from ident or meta data to use.}

\item{plot_title}{What it says on the tin.  Default: NULL}

\item{assay_use}{Assay to plot.  Default: "RNA"}

\item{flip}{Should genes be displayed along the y-axis? Default: TRUE}
}
\description{
Produce grouped violin plots.  Suitable for
plotting cluster or cell type specific marker genes.
}
