% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViolinEnsemble.R
\name{ViolinEnsemble}
\alias{ViolinEnsemble}
\title{ViolinEnsemble}
\usage{
ViolinEnsemble(
  object,
  features = NULL,
  marker_list = NULL,
  grouping_var = NULL,
  cluster_x = FALSE,
  cluster_y = FALSE,
  show_points = FALSE,
  alpha = 0.5,
  pt_size = 0.25,
  plot_title = NULL,
  assay_use = "RNA",
  slot_use = "data",
  sort = FALSE,
  flip = TRUE
)
}
\arguments{
\item{object}{Seurat object to plot}

\item{features}{List of features to plot}

\item{marker_list}{A table of cluster markers as produced by FindMarkers()
or FindAllMarkers or a two column data.frame with the columns "cluster" and
"feature".  If provided, supercedes the features argument.  Note: if a marker is
present in multiple clusters, only the first (as determined by arrange()) is used.
Default: NULL}

\item{grouping_var}{Grouping variable from ident or meta data to use. If NULL,
the current active.ident is used.  Default: NULL}

\item{cluster_x}{Arrange the x-axis variables using hierarchical clustering. Default: TRUE}

\item{cluster_y}{Arrange the y-axis variables using hierarchical clustering. Default: FALSE}

\item{show_points}{Display data points in addition to violin using geom_jitter? Default: FALSE}

\item{alpha}{Alpha value to use with geom_jitter. Default: 1}

\item{pt_size}{Point size to use with geom_jitter. Default: 0.25}

\item{plot_title}{What it says on the tin.  Default: NULL}

\item{assay_use}{Assay to plot.  Default: "RNA"}

\item{slot_use}{Slot to draw data from.  Default: "data"}

\item{sort}{If TRUE, attempt to sort the groups alpha-numerically. Default: FALSE}

\item{flip}{Should features be displayed along the y-axis? Default: TRUE}
}
\value{

}
\description{
Produce grouped violin plots.  Suitable for
plotting cluster or cell type specific marker features.
}
